requirements.txt
pyTelegramBotAPI==4.14.0
aiogram-utils==0.1.4
mysql-connector-python==8.3.0
python-dotenv==1.0.0
requests==2.31.0
aiogram~=2.25.2
tgcrypto==1.2.5
aiohttp~=3.8.6
Pyrogram~=2.0.106